Move parse_composable_node to ComposableNode class
