+++ launch_ros/launch_ros/actions/composable_node_container.py
@@ -27,7 +27,6 @@ from launch.some_substitutions_type import SomeSubstitutionsType
 from .node import Node
 
 from ..descriptions import ComposableNode
-from ..utilities.parse_composable_node import parse_composable_node
 
 
 @expose_action('node_container')
@@ -64,7 +63,8 @@ class ComposableNodeContainer(Node):
         composable_nodes = entity.get_attr(
             'composable_node', data_type=List[Entity], optional=True)
         if composable_nodes is not None:
-            parsed_tuples = [parse_composable_node(parser, entity) for entity in composable_nodes]
+            parsed_tuples = [ComposableNode.parse_composable_node(
+                parser, entity) for entity in composable_nodes]
             kwargs['composable_node_descriptions'] = [t[0](**t[1]) for t in parsed_tuples]
 
         return cls, kwargs
