+++ launch_ros/launch_ros/actions/load_composable_nodes.py
@@ -47,7 +47,6 @@ from ..utilities import make_namespace_absolute
 from ..utilities import prefix_namespace
 from ..utilities import to_parameters_list
 from ..utilities.normalize_parameters import normalize_parameter_dict
-from ..utilities.parse_composable_node import parse_composable_node
 
 
 @expose_action('load_composable_node')
@@ -97,7 +96,8 @@ class LoadComposableNodes(Action):
 
         composable_nodes = entity.get_attr('composable_node', data_type=List[Entity])
         if composable_nodes is not None:
-            parsed_tuples = [parse_composable_node(parser, entity) for entity in composable_nodes]
+            parsed_tuples = [ComposableNode.parse_composable_node(
+                parser, entity) for entity in composable_nodes]
             kwargs['composable_node_descriptions'] = [t[0](**t[1]) for t in parsed_tuples]
 
         return cls, kwargs
